style: move css into styles.js
